requirements_ui.txt
flask>=3.0
pywebview>=4.0
